[clang] Improve -Wbitwise-instead-of-logical message
Some initial feedback from Linux kernel developers is that
-Wbitwise-instead-of-logical's note might be taken as the suggested fix
instead of the fix it hint between the warning and the note.

To improve this, add "did you mean logical '<op>'?" to the main warning
text to get the developer thinking about whether or not they truly
intended a bitwise operation, instead of focusing on the note to silence
instead of truly fix the warning.